requirements.txt
a2wsgi==1.10.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.2.0
apache-airflow==2.5.0
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==3.3.2
argcomplete==3.2.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachelib==0.9.0
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==3.0.5
crashtest==0.4.1
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==42.0.2
dask==2024.2.0
dask-expr==0.4.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
dnspython==2.5.0
docutils==0.20.1
dulwich==0.21.7
email-validator==1.3.1
exceptiongroup==1.2.0
Faker==23.1.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==5.13.0
inflection==0.5.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
json-merge-patch==0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.0
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
locket==1.0.0
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.1.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-proto==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
partd==1.4.1
pathspec==0.9.0
pendulum==3.0.0
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prison==0.2.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
psycopg2==2.9.9
ptyprocess==0.7.0
pyarrow==15.0.0
pyasn1==0.5.1
pybind11==2.11.1
pycparser==2.21
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-multipart==0.0.9
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.17.1
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.0
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.32.0.post1
swagger_ui_bundle==1.1.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.13.0
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
trove-classifiers==2024.1.31
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
xattr==0.10.1
yarl==1.9.4
zipp==3.17.0
